sasl:   add _c_int to buildin type.   some bugs fixed.   add operator overload.   add simple check for sasl_ehelper.   Issue 41 updated.

+++ sasl/src/code_generator/llvm/cgllvm_impl.cpp
@@ -225,7 +225,7 @@ void llvm_code_generator::visit( function_type& v ){
 
 	// Register parameter names.
 	llvm::Function::arg_iterator arg_it = fctxt->func->arg_begin();
-	for( int arg_idx = 0; arg_idx < fctxt->func->arg_size(); ++arg_idx, ++arg_it){
+	for( size_t arg_idx = 0; arg_idx < fctxt->func->arg_size(); ++arg_idx, ++arg_it){
 		boost::shared_ptr<parameter> par = v.params[arg_idx];
 		arg_it->setName( par->symbol()->unmangled_name() );
 		common_ctxt_handle par_ctxt = get_common_ctxt( par->param_type );
